update the engine api; prepare to del mini_eval
