+++ pyanomaly/core/engine/functions/anopred.py
@@ -19,7 +19,7 @@ torch.autograd.set_detect_anomaly(True)
 
 from pyanomaly.core.utils import AverageMeter, flow_batch_estimate, tensorboard_vis_images, make_info_message, ParamSet
 from pyanomaly.datatools.evaluate.utils import psnr_error
-from pyanomaly.core.engine.default_engine import DefaultTrainer, DefaultInference
+from ..abstract.default_engine import DefaultTrainer, DefaultInference
 
 from ..engine_registry import ENGINE_REGISTRY
