+++ pyanomaly/core/engine/functions/stae.py
@@ -21,7 +21,7 @@ from torch.utils.data import DataLoader
 from pyanomaly.core.utils import AverageMeter, flow_batch_estimate, tensorboard_vis_images, make_info_message, ParamSet
 from pyanomaly.datatools.evaluate.utils import psnr_error
 from pyanomaly.utils.flow_utils import flow2img
-from pyanomaly.core.engine.default_engine import DefaultTrainer, DefaultInference
+from ..abstract.default_engine import DefaultTrainer, DefaultInference
 from ..engine_registry import ENGINE_REGISTRY
 
 __all__ = ['STAETrainer', 'STAEInference']
@@ -67,8 +67,8 @@ class STAETrainer(DefaultTrainer):
         # basic meter
         self.loss_meter_STAE = AverageMeter(name='loss_STAE')
 
-        self.test_dataset_keys = self.kwargs['test_dataset_keys']
-        self.test_dataset_dict = self.kwargs['test_dataset_dict']
+        # self.test_dataset_keys = self.kwargs['test_dataset_keys']
+        # self.test_dataset_dict = self.kwargs['test_dataset_dict']
 
     def train(self,current_step):
         # Pytorch [N, C, D, H, W]
@@ -139,21 +139,22 @@ class STAETrainer(DefaultTrainer):
         self.kwargs['writer_dict']['global_steps_{}'.format(self.kwargs['model_type'])] = global_steps
     
     def mini_eval(self, current_step):
-        if current_step % self.steps.param['mini_eval'] != 0:
-            return
-        temp_meter_rec = AverageMeter()
-        # temp_meter_pred = AverageMeter()
-        self.set_requires_grad(self.STAE, False)
-        self.STAE.eval()
-        for data, _ in self.val_dataloader:
-            input_mini = data.cuda()
-            # Use the model, get the output
-            output_rec_mini, output_pred_mini = self.STAE(input_mini)
-            rec_psnr_mini = psnr_error(output_rec_mini.detach(), input_mini)
-            # pred_psnr_mini = psnr_error(output_pred_mini.detach(), input_pred_mini)
-            temp_meter_rec.update(rec_psnr_mini.detach())
-            # temp_meter_pred.update(pred_psnr_mini.detach())
-        self.logger.info(f'&^*_*^& ==> Step:{current_step}/{self.steps.param["max"]} the REC PSNR is {temp_meter_rec.avg:.3f}')
+        # if current_step % self.steps.param['mini_eval'] != 0:
+        #     return
+        # temp_meter_rec = AverageMeter()
+        # # temp_meter_pred = AverageMeter()
+        # self.set_requires_grad(self.STAE, False)
+        # self.STAE.eval()
+        # for data, _ in self.val_dataloader:
+        #     input_mini = data.cuda()
+        #     # Use the model, get the output
+        #     output_rec_mini, output_pred_mini = self.STAE(input_mini)
+        #     rec_psnr_mini = psnr_error(output_rec_mini.detach(), input_mini)
+        #     # pred_psnr_mini = psnr_error(output_pred_mini.detach(), input_pred_mini)
+        #     temp_meter_rec.update(rec_psnr_mini.detach())
+        #     # temp_meter_pred.update(pred_psnr_mini.detach())
+        # self.logger.info(f'&^*_*^& ==> Step:{current_step}/{self.steps.param["max"]} the REC PSNR is {temp_meter_rec.avg:.3f}')
+        pass
 
 @ENGINE_REGISTRY.register()
 class STAEInference(DefaultInference):
