--- pyanomaly/datatools/augment/__init__.py
@@ -1,1 +1,0 @@
-from .augment_builder import AugmentAPI
